extract `Intro` block for `/partnership` page
